refactor: часть логики работы с рейтингами вынесена в отдельный сервис.
